updated for tests4j

+++ src/org/adligo/tests4j_4gwt/client/ui/I_TrialRunUi.java
@@ -13,7 +13,7 @@ public interface I_TrialRunUi {
 	
 	public void showTestWidget(Object o);
 	
-	public void setTotalTests(int p);
+	public void setTotalTests(long p);
 	
 	public void setRunHandler(I_RunHandler handler);
